Luna requirements.txt
SpeechRecognition
pyttsx3
pywhatkit
wikipedia
pyjokes
requests
beautifulsoup4
googlesearch-python
pandas
openpyxl
fuzzywuzzy
python-Levenshtein
termcolor
